+++ .gitignore
@@ -1,5 +1,6 @@
 node_modules
 out
+out-web
 .vscode-test
 *.code-workspace
 */*.tsbuildinfo
