+++ client/package.json
@@ -8,9 +8,13 @@
 	},
 	"scripts": {
 		"esbuild": "npx esbuild ./src/main.ts --bundle --outfile=out/main.js --external:vscode --format=cjs --platform=node",
+		"esbuild-web": "npx esbuild ./src/main-web.ts --bundle --outfile=out-web/main.js --external:vscode --format=cjs --platform=browser",
 		"build:prod": "npm run esbuild -- --minify",
 		"build": "npm run esbuild -- --sourcemap",
-		"watch": "npm run esbuild -- --sourcemap --watch=forever"
+		"watch": "npm run esbuild -- --sourcemap --watch=forever",
+		"build-web:prod": "npm run esbuild-web -- --minify",
+		"build-web": "npm run esbuild-web -- --sourcemap",
+		"watch-web": "npm run esbuild-web -- --sourcemap --watch=forever"
 	},
 	"devDependencies": {
 		"@types/semver": "^7.5.6",
@@ -18,6 +22,6 @@
 	},
 	"dependencies": {
 		"semver": "^7.5.4",
-		"vscode-languageclient": "^8.1.0"
+		"vscode-languageclient": "^9.0.1"
 	}
 }
